requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Computer Vision (lightweight)
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=8.3.2,<10.1.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
